requirements.txt
asgiref==3.2.10
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==38.0.4
defusedxml==0.7.0rc1
Deprecated==1.2.13
Django==3.0.8
django-allauth==0.52.0
django-cors-headers==3.7.0
django-oauth-toolkit==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.5.0
drf-social-oauth2==1.2.1
ecdsa==0.18.0
idna==2.10
jwcrypto==1.4.2
oauthlib==3.1.0
Pillow==9.3.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
python-jose==3.3.0
python3-openid==3.2.0
pytz==2020.1
requests==2.25.0
requests-oauthlib==1.3.0
rsa==4.9
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.3.2.post1
sqlparse==0.3.1
urllib3==1.26.2
wrapt==1.14.1
